requirements.txt
azure-identity
azure-mgmt-network
boto3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-sqlalchemy
pymysql
google-api-python-client
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kubernetes
MarkupSafe==1.1.1
oauth2client
pystache==0.5.4
python-dotenv==0.17.1
pyyaml
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
